Have modified pages_controller_spec to use the ApplicationHelper.title method rather than repeating the huge "this and such is my base title" strings.
Modified ApplicationHelper.title to take an optional argument so that the pages_controller_spec can say ":content => title 'Home'" (or something like that), rather than repeating "big ass base title".

Users controller now makes sure that if the user is signed in, it will redirect to the home page if the user tries new or create.

This is all part of exercise 3 in chapter 10 of the tutorial book.