working to get some styling done.  css is a real struggler for me. I used sass inheritance to make the footer have an id of footer_nav which inherits from site_nav, so I could override the font size of the site_nav ul li item.  Otherwise, using "footer nav ul li a" made the font smaller, but left the li border too large.
my other efforts made the font smaller, but the border *way* to short.

Now it looks nice.

Did a gradient for the header, not sure how much you can tell.

made the site/footer nav have a gradient and a box shadow.
Had been thinking about making the ul round and white background.

made the body main have a box shadow and a dark backround.
made the text gray.  did a bit to center it, but it's not working.